통신 에러 처리, 텍스트뷰 익스텐션 추가

+++ DMZing/DMZing/Course/Model/CourseDetail.swift
@@ -12,7 +12,7 @@ typealias CourseDetailData = CourseDetail
 
 struct CourseDetail: Codable {
     let backgroundImageUrl: String 
-    let estimatedTime: Int
+    let estimatedTime: Double
     let id: Int
     let imageUrl: String
     let level: String
